refactor: 클래스 명 변경
변경 전: GameResultMaker
변경 후: Circuit
변경 이유: 자동차 경주 게임을 진행하는 클래스임을 알리기 위함
